TIP-871: Move comment ui stuff to Akeneo\Enrichment
